Moving Task into droidlet/shared_data_structs.py
Moving get_default_conditions to shared_data_structs
