+++ examples_and_tutorials/grid/tasks.py
@@ -1,5 +1,5 @@
 import numpy as np
-from droidlet.interpreter.task import Task
+from droidlet.shared_data_structs import Task
 
 
 class Move(Task):
